fix: split share function
